requirements.txt
efficientnet_pytorch==0.7.1
einops==0.3.0
imgaug==0.4.0
joblib==0.14.1
matplotlib==3.1.3
numpy==1.18.1
opencv_python==4.2.0.32
pandas==1.1.4
Pillow==9.1.0
PyYAML==6.0
scikit_learn==1.0.2
scipy==1.4.1
timm==0.5.4
torch==1.4.0
torchvision==0.5.0
tqdm==4.61.1

